middleware: extract shared-athlete conflict-edge builders to scheduling/conflict_edges.py (#152 step 7)
5 methods moved verbatim: core gym team lookup, gym team conflicts,
team-state-to-unit normalization, shared-athlete edge construction, and
cross-sport conflict assembly. Pod doubles/singles resolution stays in
the facade (shared with the planning review tabs). No behavior changes.
